requirements.txt
pandas
neo4j 
openai
python-dotenv
openai
pinecone
sqlite-utils
sentence_transformers
nltk
rank-bm25
langchain_core
langchain_openai
redis
langchain
fuzzywuzzy 
python-Levenshtein
langgraph
langchain-community
fastapi
langsmith
uvicorn
python-dotenv